src/ml_service/requirements.txt
grpcio
grpcio-tools
faiss-cpu
sentence-transformers
